API仕様と合わせてステータスコードの修正
